LipidTrend {LipidTrend}	R Documentation

LipidTrend: Analysis and Visualization of Lipid Feature Tendencies

Description

The LipidTrend package provides tools for analyzing and visualizing trends in lipidomics data. It implements statistical methods for identifying significant changes in lipid abundances across different features, with support for both one-dimensional and two-dimensional analyses.

Main functions

analyzeLipidRegion: Analyzing lipid trends using permutation tests and Gaussian kernel smoothing

plotRegion1D: Visualize one-dimensional lipid trends

plotRegion2D: Create two-dimensional visualizations of lipid trends

Vignettes

See the package vignettes for detailed workflows: vignette('LipidTrend')

Installation

To install from Bioconductor, use:

if (!requireNamespace('BiocManager', quietly=TRUE))
    install.packages('BiocManager')
BiocManager::install('LipidTrend')

Author(s)

Maintainer: Wei-Chung Cheng chenglabomics@gmail.com (ORCID) [copyright holder]

Authors:

Chia-Hsin Liu [contributor]

Pei-Chun Shen [contributor]

Wen-Jen Lin [contributor]

Hung-Ching Chang [contributor]

Meng-Hsin Tsai [contributor]

See Also

Useful links:

https://github.com/BioinfOMICS/LipidTrend

Report bugs at https://github.com/BioinfOMICS/LipidTrend/issues

[Package LipidTrend version 1.0.0 Index]