char_table_2D {LipidTrend}	R Documentation

Example lipid characteristics table for two-dimensional LipidTrend analysis

Description

Example lipid characteristics table for two-dimensional LipidTrend analysis

Usage

data(char_table_2D)

Format

An object of class data.frame with 137 rows and 2 columns.

Value

A data.frame object of total chain length and total double bond characteristics over 137 lipids

Source

Tomoyuki Shiota et al. ,Hepatoviruses promote very-long-chain fatty acid and sphingolipid synthesis for viral RNA replication and quasi-enveloped virus release. Sci Adv. 9(42):eadj4198 https://www.science.org/doi/10.1126/sciadv.adj4198.

Examples

data(char_table_2D)

[Package LipidTrend version 1.0.0 Index]