even_chain_result {LipidTrend}	R Documentation

Get Even Chain Result from LipidTrendSE

Description

Get Even Chain Result from LipidTrendSE

Usage

even_chain_result(object)

## S4 method for signature 'LipidTrendSE'
even_chain_result(object)

Arguments

object
A LipidTrendSE object

Value

A data frame containing even chain result. The result table includes the following columns:

Feature columns: Lipid feature values from the input rowData, such as chain length or double bond count. Column names vary depending on input.

avg.abund: Mean abundance of each lipid across all samples. For one-dimensional analysis, this may also include avg.abund.ctrl and avg.abund.case for group-wise means.

direction: Sign of the smoothed statistic:

+ : Trend increases in the case group.

- : Trend decreases in the case group.

smoothing.pval.BH: Benjamini–Hochberg adjusted p-value from the region-based permutation test.

marginal.pval.BH: Benjamini–Hochberg adjusted p-value from the marginal test (per lipid).

log2.FC: Log2 fold-change in abundance between case and control groups.

significance: Overall significance label based on smoothed test and FC direction:

Increase: Significant positive trend in case group.

Decrease: Significant negative trend in case group.

NS: Not significant.

Examples

data("lipid_se_CL")
sub <- lipid_se_CL[seq_len(10), ]
res_se <- analyzeLipidRegion(
    lipid_se=sub, ref_group="sgCtrl", split_chain=TRUE,
    chain_col="chain", radius=3, own_contri=0.5, permute_time=100)
# Get complete result summary
results <- even_chain_result(res_se)

[Package LipidTrend version 1.0.0 Index]