plotRegion1D {LipidTrend}	R Documentation

Plot region trends for one-dimensional lipid feature

Description

Visualize lipid trends from one-dimensional analysis results. The plot highlights regions of significant group differences based on the smoothed permutation test. Blue and red ribbons mark where the trend significantly decreases or increases, respectively, in the case group compared to the control group. Each point represents the mean abundance within each group (case vs. control) for a specific lipid feature. This visualization helps identify not only abundance differences but also the specific lipid feature regions where trends diverge between groups.

Usage

plotRegion1D(object, p_cutoff = 0.05, y_scale = "identity")

## S4 method for signature 'LipidTrendSE'
plotRegion1D(object, p_cutoff = 0.05, y_scale = "identity")

Arguments

object
A LipidTrendSE object generated by analyzeLipidRegion() for one-dimensional feature analysis.

p_cutoff
Numeric. Significance threshold for highlighting regions based on smoothed permutation p-values. Default is 0.05.

y_scale
Character. Scale of the y-axis. Choose from "identity", "log2", "log10", or "sqrt". Default is "identity".

Value

If split_chain=TRUE in the analysis, the function returns a list containing two plots:

even_result: A ggplot object visualizing trends for even-chain lipids, or NULL if no such lipids exist.

odd_result: A ggplot object visualizing trends for odd-chain lipids, or NULL if no such lipids exist.

If split_chain=FALSE, the function returns a single ggplot object visualizing the trend for all lipids combined.

See Also

analyzeLipidRegion for generating the input LipidTrendSE object

Examples

data("lipid_se_CL")
res_se <- analyzeLipidRegion(
    lipid_se=lipid_se_CL, ref_group="sgCtrl", split_chain=FALSE,
    chain_col=NULL, radius=2, own_contri=0.5, test="t.test",
    abund_weight=TRUE, permute_time=100)
plot <- plotRegion1D(res_se, p_cutoff=0.05, y_scale='identity')

[Package LipidTrend version 1.0.0 Index]